Change promtail port and add 514/tcp for syslog-ng
Promtail only supports RFC5424 on syslog input, the standard port for
it is 601/tcp, so we change it for 1601/tcp for a better match.
Some equipments that use RFC3164 "BSD-syslog" can be configured to use
TCP instead of of the standard UDP so add a syslog-ng TCP listener on
port 514 for RFC3164 formatted messages.